+++ atuin-server/src/handlers/v0/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod record;
